[confiscan] Gather users and group information
This commit gathers basic user and group information. Everything is
retrieved from `/etc/passwd` and `/etc/groups` and then parsed to a CSV
format which is written to the output directory.

All previliged users are listed by checking whether they are part of the
`wheel` or `sudo` group. The result is not written to a file, since This
can be derived from the parsed passwd output.