[confiscan] Gather application-specific configs
Configuration files of installed application can now be specified by
adding the path to the end of the argument list, no need for specifying
an option. This ensures that Shell globbing is supported for
[CONFIG_FILES], and can be a file or directory.

For example, `./confiscan.sh /etc/sudoers /etc/sysctl.conf` will
result in the following structure:

```
srv5-configs/
├── etc
│   ├── sudoers
│   └── sysctl.conf
└── ...
```

When no path is specified, the whole `/etc/` directory will be used
instead.