[poc] Increase memory from srv3
During testing of the poc.yml playbook, Ansible consistently failed
with the following error:

```
fatal: [srv3]: FAILED! => {
    "changed": false,
    "module_stderr": "Shared connection to 172.16.128.3 closed.\r\n",
    "module_stdout": "Killed\r\n",
    "msg": "MODULE FAILURE\nSee stdout/stderr for the exact error",
    "rc": 137
}
```

Upon investigation of `/var/log/syslog`, it became evident that srv3 was
experiencing out-of-memory issues consistently during the execution of
the copy task. To preemptively resolve this issue, I increased the
memory allocation of srv3 to 2GB, as the server was consistently
utilizing around 1.2GB of memory during task execution.