[confiscan] Fix broken links after copy
This commit addresses a problem related to copying symlinks.
When a symbolic link was copied, and it pointed to a location not
present in the specified config paths, it will break the link,
making it unavailable in the inventory. With `rsync` we can follow
these pointers and copy the actual file.